Various type fixes found when compiling with -Wconversion. Delete some unused functions and convert some graphics macros to static inline functions.

+++ src/a2x_pack_bitfield.c
@@ -20,7 +20,7 @@
 #include "a2x_pack_bitfield.v.h"
 
 typedef unsigned long AChunk;
-#define BITS_PER_CHUNK (sizeof(AChunk) * 8)
+#define BITS_PER_CHUNK (unsigned)(sizeof(AChunk) * 8)
 #define BITS_PER_CHUNK_MASK (BITS_PER_CHUNK - 1)
 
 struct ABitfield {
@@ -50,13 +50,13 @@ void a_bitfield_free(ABitfield* Bitfield)
 
 void a_bitfield_set(ABitfield* Bitfield, unsigned Bit)
 {
-    const AChunk bit = 1 << (Bit & BITS_PER_CHUNK_MASK);
+    const AChunk bit = 1u << (Bit & BITS_PER_CHUNK_MASK);
     Bitfield->bits[Bit / BITS_PER_CHUNK] |= bit;
 }
 
 void a_bitfield_clear(ABitfield* Bitfield, unsigned Bit)
 {
-    const AChunk bit = 1 << (Bit & BITS_PER_CHUNK_MASK);
+    const AChunk bit = 1u << (Bit & BITS_PER_CHUNK_MASK);
     Bitfield->bits[Bit / BITS_PER_CHUNK] &= ~bit;
 }
 
@@ -68,7 +68,7 @@ void a_bitfield_reset(ABitfield* Bitfield)
 bool a_bitfield_test(const ABitfield* Bitfield, unsigned Bit)
 {
     const AChunk value = Bitfield->bits[Bit / BITS_PER_CHUNK];
-    const AChunk bit = 1 << (Bit & BITS_PER_CHUNK_MASK);
+    const AChunk bit = 1u << (Bit & BITS_PER_CHUNK_MASK);
 
     return (value & bit) != 0;
 }
